# specs/001-asset-ledger-demo/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Assets API
  version: 0.1.0
  description: >
    内部路由设备资产台账与流转系统。全局约定（认证、错误 envelope、分页、乐观锁）见 README.md。
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /assets:
    get:
      tags: [assets]
      summary: 资产列表。搜索精确优先、否则子串；支持任意 attr.<key> 等值筛选
      parameters:
        - { name: q, in: query, schema: { type: string }, description: 精确匹配任一唯一字段的当前值或历史值，其次 UUID；未命中则子串匹配 }
        - { name: category_id, in: query, schema: { type: string } }
        - { name: include_descendants, in: query, schema: { type: boolean, default: true } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/AssetStatus" } }
        - { name: owner_id, in: query, schema: { type: string } }
        - { name: holder_type, in: query, schema: { type: string, enum: [user, entity] } }
        - { name: holder_id, in: query, schema: { type: string } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Asset" }
                      exact_match_id:
                        type: string
                        nullable: true
                        description: q 唯一精确命中时给出，前端据此直达详情页
    post:
      tags: [assets]
      summary: 创建资产。跑完整保存管线并写 kind=create 流转事件
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetWrite" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /assets/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [assets]
      summary: 资产详情。含已归档信息项的取值
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [assets]
      summary: 修改属性或变更类别。必须带 version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/AssetWrite"
                - type: object
                  required: [version]
                  properties:
                    version: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [assets]
      summary: 硬删并级联流转历史与唯一值。需带 confirm 二次确认（输入该资产的显示编号）
      parameters:
        - { name: confirm, in: query, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "422": { $ref: "#/components/responses/Error" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: 统一错误 envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 英文 snake_case }
            message: { type: string, description: 中文，用户可见 }
            fields:
              type: object
              additionalProperties: { type: string }
    Page:
      type: object
      required: [items, total, offset, limit]
      properties:
        total: { type: integer }
        offset: { type: integer }
        limit: { type: integer }
    AssetStatus:
      type: string
      enum: [in_stock, in_use, in_repair, lost, retired]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        auth_type: { type: string, enum: [oidc, local] }
        status: { type: string, enum: [active, disabled] }
    HolderRef:
      type: object
      required: [type, id]
      properties:
        type: { type: string, enum: [user, entity] }
        id: { type: string }
        name: { type: string, readOnly: true }
        entity_type:
          type: string
          nullable: true
          description: type=entity 时给出 company / location / department
    Asset:
      type: object
      properties:
        id: { type: string, description: UUID，唯一内部标识 }
        display_name: { type: string, description: 由类别 display_key 求值得到；未配置时为 UUID 前 8 位 }
        category_id: { type: string }
        model_id: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/AssetStatus" }
        owner: { $ref: "#/components/schemas/User" }
        holder: { $ref: "#/components/schemas/HolderRef" }
        attrs:
          type: object
          additionalProperties: true
        archived_attrs:
          type: object
          description: 不属于当前有效信息项集的孤儿键，只读展示
          additionalProperties: true
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AssetWrite:
      type: object
      required: [category_id, owner_id, holder_type, holder_id]
      properties:
        category_id: { type: string }
        model_id: { type: string, nullable: true }
        owner_id: { type: string }
        holder_type: { type: string, enum: [user, entity] }
        holder_id: { type: string }
        status: { $ref: "#/components/schemas/AssetStatus" }
        attrs:
          type: object
          description: 不含 computed 项，由服务端求值
          additionalProperties: true
...